Weaker form of staged locks/invariants, proof of shadow ptr to show use.
This avoids the need for leases is the shadow pointer atomic pair proof, and
I think it makes it quite a bit simpler. It is also shorter:

     spec    proof comments
      119      154        2 RefinementShadowStaged.v
      142      198        9 RefinementShadow.v

(Some of the savings is because after making the change, it was clearer to me
how to shorten the simplified proof further).

This is not much longer than the POCS AtomicPair reference solution now.